requirements.txt
beautifulsoup4==4.3.2
cssselect==0.9.1
feedfinder2==0.0.1
feedparser==5.1.3
ipython==3.2.0
jieba3k==0.35.1
lxml==3.3.5
newspaper3k==0.1.5
nltk==3.0.1
Pillow==2.6.1
python-dateutil==2.4.0
PyYAML==3.11
requests==2.3.0
six==1.7.3
tldextract==1.5.1
wheel==0.24.0
